Implement lifespan management and background storm stopping; refactor NewStorm components to use context for state management
