+++ UI/src/context/CustomMUIPropsContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+const CustomMUIPropsContext = createContext();
