+++ UI/src/context/SystemInfoContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+const SystemInfoContext = createContext();
